requirements.txt
botbuilder-integration-aiohttp>=4.8.0
jira
docxtpl
yadisk
python-dotenv